chore: reorganize directives like components
