+++ packages/chusho/src/directives/clickOutside/index.ts
@@ -1,0 +1,3 @@
+import clickOutside from './clickOutside';
+
+export { clickOutside };
