--- packages/chusho/src/utils/uuid.spec.js
@@ -1,9 +1,0 @@
-import uuid from './uuid';
-
-describe('uuid', () => {
-  it('should automatically increments', () => {
-    expect(uuid()).toEqual('0');
-    expect(uuid()).toEqual('1');
-    expect(uuid('prefix')).toEqual('prefix-2');
-  });
-});
